Get zone when needed, name from GetInstanceInfo.
Instead of keeping track of the current zone, we now ask for it when
we need it. Probably doesn't make much of a difference either way, but
tracking it was just complexity for nothing.

Also changed to using the name returned by GetInstanceInfo. This seems
to be the same as we got previously from GetRealZoneText anyway(?)