Options: Separate new tracked raids from old
Breaks the list of tracked raids into two, one for Current Expansion and
one for Old Expansions. The separation is made by looking up a hardcoded
list of raid instance ids. If the raid instance id is known to belong to
an old raid it is placed in the Old Expansions list, otherwise it is
seen as belonging to the Current Expansion.